refactor: Hoist host clipboard write-back into HostClipboardPublisher
## Summary
- Extract the "Copy to Mac" host-pasteboard write path out of the
  window-scoped ClipboardContentViewController into a new window-independent
  HostClipboardPublisher, so automatic clipboard passthrough (#82) can publish
  guest content to the host pasteboard with the clipboard window closed.
- No behavior change: the window's Copy-to-Mac button delegates to the publisher
  and maps the result to the same status messages.

## Changes
- Add Kernova/Services/HostClipboardPublisher.swift owning the pasteboard-item
  planner (hostPasteboardItems / lazyFileSpec / stagedFileURL / inlineData), the
  host-write staging, and provider-registry retention; publish(from:) returns a
  HostPublishOutcome the window maps to transient messages.
- Slim ClipboardContentViewController: copyToMac now awaits publisher.publish and
  calls showCopyOutcome; its own ClipboardFileStaging is retained only for the
  Paste/drop folder-archiving intake (distinct from host-write staging, sharing
  the launch-swept "host" root without collision).
- Add changeCount to the HostWritePasteboard seam (post-write change count feeds
  the upcoming passthrough echo suppression); NSPasteboard already satisfies it,
  the test fake mirrors it.
- Move the shared staging label to HostClipboardPublisher.stagingLabel (AppDelegate
  launch sweep updated) and repoint the host-pasteboard-items tests.

## Test plan
- [x] make build
- [x] ClipboardContentViewControllerRetentionTests pass (Copy-to-Mac write/retain/fail)
- [x] ClipboardHostPasteboardItemsTests pass (grouping/staging via new home)
- [x] swift-format --strict clean on changed files
